requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
openai==1.40.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
tqdm==4.66.5
pydub==0.25.1
vosk==0.3.44
aiofiles==23.2.1
ruff==0.5.7
black==24.8.0
aiosqlite==0.20.0
